requirements.txt
beautifulsoup4==4.14.3
chromadb==0.5.3
fastapi==0.135.0
uvicorn==0.30.1
groq==1.0.0
pandas==2.2.2
pydantic==2.12.5
pysqlite3==0.6.0
python-dotenv==1.2.2
requests==2.32.5
streamlit==1.45.1
sentence-transformers==2.7.0